// src/types/neo4j/nodes.ts
export interface BaseNodeInterface {
    w: number;
    h: number;
    color: string;
    __primarylabel__: string;
    unique_id: string;
    path: string;
    created: string;
    merged: string;
}

// Users
export interface UserNodeInterface extends BaseNodeInterface
{
    user_id: string;
    user_type: string;
    user_name: string;
    user_email: string;
    worker_node_data: string;
}

export interface DeveloperNodeInterface extends BaseNodeInterface{
    user_id: string;
    user_type: string;
    user_name: string;
    user_email: string;
}

export interface TeacherNodeInterface extends BaseNodeInterface
{
    teacher_code: string;
    teacher_name_formal: string;
    teacher_email: string;
    worker_db_name: string;
}